# backend/openapi.yaml
openapi: 3.0.3
info:
  title: CourseForge API
  description: |
    REST API for the CourseForge SPA workflow:

    **Brief → Plan → PPT → Lab → Lab guide → Complete**

    Most mutating workflow endpoints return the updated course immediately and
    continue work in a background job. Poll `GET /courses/{courseId}` until
    status leaves `*_GENERATING`.

    ## Standalone Agent Microservices
    External agent endpoints configured via environment variables:
    - **PPTX Agent** (`AGENT_PPTX_URL`): Generates `.pptx` presentation binaries
    - **Lab Agent** (`AGENT_LAB_URL`): Generates hands-on lab exercises and tasks
    - **Lab Guide Agent** (`AGENT_LAB_GUIDE_URL`): Generates structured walkthrough documentation

    ## Auth
    Use **Authorize** with a JWT from `POST /auth/login` as:
    `Bearer <token>`

    Demo instructor: `instructor@eduservit.local` / `CourseForge123!`
  version: 1.1.0
  contact:
    name: EduservIT
servers:
  - url: http://127.0.0.1:8080
    description: Local development
  - url: http://localhost:8080
    description: Local development (localhost)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from `/auth/login` or `/auth/signup`.
  parameters:
    CourseId:
      name: courseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Course ID
    ArtifactId:
      name: artifactId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SlideFilename:
      name: filename
      in: path
      required: true
      schema:
        type: string
        pattern: '^slide-\d{2}\.png$'
        example: slide-01.png
      description: Pre-rendered slide image filename
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
        code:
          type: string
          example: validation
      required: [message]
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [admin, instructor, reviewer, learner]
        permissions:
          type: array
          items:
            type: string
    Session:
      type: object
      properties:
        token:
          type: string
          description: JWT for Authorization Bearer header
        user:
          $ref: '#/components/schemas/User'
    AuthCredentials:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: instructor@eduservit.local
        password:
          type: string
          format: password
          example: CourseForge123!
        name:
          type: string
          description: Optional on signup; derived from email if omitted
...
    Artifact:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fileId:
          type: string
          format: uuid
        type:
          type: string
          example: ppt
        name:
          type: string
          example: access-token-management-theory.pptx
        label:
          type: string
        sizeLabel:
          type: string
        mimeType:
          type: string
        downloadUrl:
          nullable: true
          type: string
...
    ArtifactList:
      type: object
      properties:
        artifacts:
          type: array
          items:
            $ref: '#/components/schemas/Artifact'
...
paths:
...
  /auth/signup:
    post:
      tags: [Auth]
      summary: Create an instructor account
      operationId: signup
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthCredentials'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags: [Auth]
      summary: Login and receive JWT
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthCredentials'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout (client should discard token)
      operationId: logout
      security:
        - BearerAuth: []
      responses:
        '200':
          description: OK
  /auth/me:
    get:
      tags: [Auth]
      summary: Current user profile
      operationId: me
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
...
  /courses/{courseId}/ppt:
    parameters:
      - $ref: '#/components/parameters/CourseId'
    get:
      tags: [Presentation]
      summary: Stream the course PPTX presentation file directly for in-browser rendering
      description: Returns the raw binary .pptx presentation file for vector rendering via pptx-viewer.
      operationId: getCoursePpt
      security:
        - BearerAuth: []
      responses:
        '200':
          description: PPTX presentation binary stream
          content:
            application/vnd.openxmlformats-officedocument.presentationml.presentation:
              schema:
                type: string
                format: binary
        '404':
          description: PPT presentation not found
...
  /courses/{courseId}/artifacts:
    parameters:
      - $ref: '#/components/parameters/CourseId'
    get:
      tags: [Artifacts]
      summary: List course artifacts
      operationId: listArtifacts
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Artifact list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactList'
  /courses/{courseId}/artifacts/{artifactId}/download:
    parameters:
      - $ref: '#/components/parameters/CourseId'
      - $ref: '#/components/parameters/ArtifactId'
    get:
      tags: [Artifacts]
      summary: Download an artifact file
      description: Returns the binary file (e.g. `.pptx`) with Content-Disposition attachment.
      operationId: downloadArtifact
      security:
        - BearerAuth: []
      responses:
        '200':
          description: File bytes
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
            application/vnd.openxmlformats-officedocument.presentationml.presentation:
              schema:
                type: string
                format: binary
        '404':
          description: Missing artifact
  /courses/{courseId}/slides/{filename}:
    parameters:
      - $ref: '#/components/parameters/CourseId'
      - $ref: '#/components/parameters/SlideFilename'
    get:
      tags: [Presentation]
      summary: Get a pre-rendered slide preview PNG
      description: Used by the Presentation screen `<img>` preview (not the downloadable PPTX).
      operationId: getSlideImage
      security:
        - BearerAuth: []
      responses:
        '200':
          description: PNG image
          content:
            image/png:
              schema:
                type: string
                format: binary
        '404':
          description: Slide image not found